+++ pandas/core/api.py
@@ -2,6 +2,17 @@
 
 import numpy as np
 
+from pandas.core.dtypes.dtypes import (
+    CategoricalDtype,
+    DatetimeTZDtype,
+    IntervalDtype,
+    PeriodDtype,
+)
+from pandas.core.dtypes.missing import isna, isnull, notna, notnull
+
+# TODO: Remove import when statsmodels updates #18264
+from pandas.core.algorithms import factorize, unique, value_counts
+from pandas.core.arrays import Categorical
 from pandas.core.arrays.integer import (
     Int8Dtype,
     Int16Dtype,
@@ -12,44 +23,33 @@ from pandas.core.arrays.integer import (
     UInt32Dtype,
     UInt64Dtype,
 )
-from pandas.core.algorithms import factorize, unique, value_counts
-from pandas.core.dtypes.missing import isna, isnull, notna, notnull
-from pandas.core.dtypes.dtypes import (
-    CategoricalDtype,
-    PeriodDtype,
-    IntervalDtype,
-    DatetimeTZDtype,
-)
-from pandas.core.arrays import Categorical, array
+from pandas.core.construction import array
+from pandas.core.frame import DataFrame
 from pandas.core.groupby import Grouper, NamedAgg
-from pandas.io.formats.format import set_eng_float_format
 from pandas.core.index import (
-    Index,
     CategoricalIndex,
-    Int64Index,
-    UInt64Index,
-    RangeIndex,
-    Float64Index,
-    MultiIndex,
-    IntervalIndex,
-    TimedeltaIndex,
     DatetimeIndex,
-    PeriodIndex,
+    Float64Index,
+    Index,
+    Int64Index,
+    IntervalIndex,
+    MultiIndex,
     NaT,
+    PeriodIndex,
+    RangeIndex,
+    TimedeltaIndex,
+    UInt64Index,
 )
+from pandas.core.indexes.datetimes import Timestamp, bdate_range, date_range
+from pandas.core.indexes.interval import Interval, interval_range
 from pandas.core.indexes.period import Period, period_range
 from pandas.core.indexes.timedeltas import Timedelta, timedelta_range
-from pandas.core.indexes.datetimes import Timestamp, date_range, bdate_range
-from pandas.core.indexes.interval import Interval, interval_range
-
-from pandas.core.series import Series
-from pandas.core.frame import DataFrame
-
-# TODO: Remove import when statsmodels updates #18264
-from pandas.core.reshape.reshape import get_dummies
-
 from pandas.core.indexing import IndexSlice
-from pandas.core.tools.numeric import to_numeric
-from pandas.tseries.offsets import DateOffset
+from pandas.core.reshape.reshape import get_dummies
+from pandas.core.series import Series
 from pandas.core.tools.datetimes import to_datetime
+from pandas.core.tools.numeric import to_numeric
 from pandas.core.tools.timedeltas import to_timedelta
+
+from pandas.io.formats.format import set_eng_float_format
+from pandas.tseries.offsets import DateOffset
